String buffer with small string optimization
This is a simple string buffer implementation with SSO. The strbuf is
a union of either a traditional strbuf struct with a malloc'ed
pointer, or a struct with an embedded buffer and the last member being
the space left in that buffer. Using space_left rather than a length
has the nice property that when the buffer is full, space_left being 0
also serves as the terminating NUL for the string.